requirements.txt
transformers
peft
datasets
sentence-transformers
faiss-cpu
fastapi
uvicorn
ipykernel
numpy
huggingface_hub
